+++ sdk_test.go
@@ -20,17 +20,17 @@ func TestMain(m *testing.M) {
 
 func TestRedactFunc(t *testing.T) {
 	t.Run("redact json", func(t *testing.T) {
-		exampleJSON, err := json.Marshal(exampleData)
+		exampleJSON, err := json.Marshal(ExampleData)
 		if err != nil {
 			t.Error(err)
 		}
-		res := redact(exampleJSON, exampleDataRedaction)
-		expected, _ := json.Marshal(exampleDataRedacted)
+		res := RedactJSON(exampleJSON, ExampleDataRedaction)
+		expected, _ := json.Marshal(ExampleDataRedacted)
 		assert.JSONEq(t, string(expected), string(res))
 	})
 
 	t.Run("redactHeaders", func(t *testing.T) {
-		result := redactHeaders(map[string][]string{
+		result := RedactHeaders(map[string][]string{
 			"Content-Type": {"application/json"},
 			"X-API-KEY":    {"test"},
 			"X-rando":      {"test 2"},
@@ -47,7 +47,7 @@ func TestOutgoingMiddleware(t *testing.T) {
 	client := &Client{
 		config: &Config{
 			RedactHeaders:      []string{"X-Api-Key", "Accept-Encoding"},
-			RedactResponseBody: exampleDataRedaction,
+			RedactResponseBody: ExampleDataRedaction,
 		},
 	}
 	var publishCalled bool
@@ -77,8 +77,8 @@ func TestOutgoingMiddleware(t *testing.T) {
 			assert.Greater(t, payload.Duration, 1000*time.Nanosecond)
 			assert.Equal(t, GoDefaultSDKType, payload.SdkType)
 
-			reqData, _ := json.Marshal(exampleData2)
-			respData, _ := json.Marshal(exampleDataRedacted)
+			reqData, _ := json.Marshal(ExampleData2)
+			respData, _ := json.Marshal(ExampleDataRedacted)
 
 			assert.Equal(t, reqData, payload.RequestBody)
 			assert.Equal(t, respData, payload.ResponseBody)
@@ -110,7 +110,7 @@ func TestOutgoingMiddleware(t *testing.T) {
 		assert.NotEmpty(t, body)
 		atHTTPClient := HTTPClient(r.Context())
 		_, _ = atHTTPClient.Get("http://localhost:3000/from-gorilla?param1=abc&param2=123")
-		jsonByte, err := json.Marshal(exampleData)
+		jsonByte, err := json.Marshal(ExampleData)
 		assert.NoError(t, err)
 
 		w.Header().Add("Content-Type", "application/json")
@@ -127,77 +127,8 @@ func TestOutgoingMiddleware(t *testing.T) {
 			"Content-Type": "application/json",
 			"X-API-KEY":    "past-3",
 		},
-		req.BodyJSON(exampleData2),
+		req.BodyJSON(ExampleData2),
 	)
 	assert.NoError(t, err)
 	assert.True(t, publishCalled)
-}
-
-var exampleData = map[string]interface{}{
-	"status": "success",
-	"data": map[string]interface{}{
-		"message": "hello world",
-		"account_data": map[string]interface{}{
-			"batch_number":            12345,
-			"account_id":              "123456789",
-			"account_name":            "test account",
-			"account_type":            "test",
-			"account_status":          "active",
-			"account_balance":         "100.00",
-			"account_currency":        "USD",
-			"account_created_at":      "2020-01-01T00:00:00Z",
-			"account_updated_at":      "2020-01-01T00:00:00Z",
-			"account_deleted_at":      "2020-01-01T00:00:00Z",
-			"possible_account_types":  []string{"test", "staging", "production"},
-			"possible_account_types2": []string{"test", "staging", "production"},
-		},
-	},
-}
-
-var exampleDataRedaction = []string{
-	"$.status", "$.data.account_data.account_type",
-	"$.data.account_data.possible_account_types",
-	"$.data.account_data.possible_account_types2[*]",
-	"$.non_existent",
-}
-
-var exampleDataRedacted = map[string]interface{}{
-	"status": "[CLIENT_REDACTED]",
-	"data": map[string]interface{}{
-		"message": "hello world",
-		"account_data": map[string]interface{}{
-			"batch_number":            12345,
-			"account_id":              "123456789",
-			"account_name":            "test account",
-			"account_type":            "[CLIENT_REDACTED]",
-			"account_status":          "active",
-			"account_balance":         "100.00",
-			"account_currency":        "USD",
-			"account_created_at":      "2020-01-01T00:00:00Z",
-			"account_updated_at":      "2020-01-01T00:00:00Z",
-			"account_deleted_at":      "2020-01-01T00:00:00Z",
-			"possible_account_types":  "[CLIENT_REDACTED]",
-			"possible_account_types2": []string{"[CLIENT_REDACTED]", "[CLIENT_REDACTED]", "[CLIENT_REDACTED]"},
-		},
-	},
-}
-
-var exampleData2 = map[string]interface{}{
-	"status": "request",
-	"send": map[string]interface{}{
-		"message": "hello world",
-		"account_data": []map[string]interface{}{{
-			"batch_number":           12345,
-			"account_id":             "123456789",
-			"account_name":           "test account",
-			"account_type":           "test",
-			"account_status":         "active",
-			"account_balance":        "100.00",
-			"account_currency":       "USD",
-			"account_created_at":     "2020-01-01T00:00:00Z",
-			"account_updated_at":     "2020-01-01T00:00:00Z",
-			"account_deleted_at":     "2020-01-01T00:00:00Z",
-			"possible_account_types": []string{"test", "staging", "production"},
-		}},
-	},
 }
